chore: Add isNewData flag to AppContext
The `isNewData` flag has been added to the `AppContext` class to track whether new data has been uploaded. This flag is set to `False` by default and is updated to `True` when new data is successfully uploaded in the `UploadFilesView` component. The flag is then reset to `False` after the data has been processed. This change allows for better tracking and handling of new data in the application.